requirements.txt
anthropic
pyboy
Pillow